+++ snippets/symbol_u32_value.c
@@ -3,6 +3,6 @@
 #include <{0}>
 
 int main() {{
-	fprintf(stdout, "%llu", (uint64_t)({1}));
+	fprintf(stdout, "%llu", (unsigned long long)(uint64_t)({1}));
 	return 0;
 }}
